[glf] Remove incorrect logic in GlfSimpleShadowArray.
For context: this code was incorrectly setting _texturesAllocatedExternally to
false in some situations. This meant that when
GlfSimpleShadowArray::_FreeResources was called, it would attempt to delete
textures it didn't actually own. In the case of the bug, the deleted texture id
had already 1) been correctly deleted in HgiGL and 2) reused in another part of
the code. This lead to a cascade of incorrect texture id usages,
and thus GL errors about invalid framebuffer operations when attempting to use
this texture as an attachment.

(Internal change: 2353146)