Fix Sample and Display Filter test failures
On sync riley versions of sample and display filters are created and stored in
HdPrman's RenderParam. Change 2347140 enabled parallel syc for these
filters, which means that when we have multiple filters, we can have multiple
threads updating RenderParam at the same time. So RenderParam's map
size is sometimes incremented however only one of of the entries is actually
in the map. This change updates the map to a concurrent_unordered_map to
avoid this issue.

(Internal change: 2354216)