Allow saving a muted SdfLayer.
There appears to be no reason not to allow saving a muted layer.
It was most likely not allowed due to the implementation, which
sets the layer's _data to an empty data object while muted.  If
we saved that we'd get an empty layer.  So now we swap the real
data in for the save and back out again if the layer is muted.

This fixes a few places that either prevented saving a muted
layer or unmuted, saved, and re-muted.  We no longer need to do
any of that.

Note that you still can't edit a muted layer.  That would be bad
given the current implementation.  An alternative implementation
would be to not compose muted layers.  That feature already exists
in Pcp but there's a significant semantic difference:  it will not
include muted layers in any layer stack, unlike the SdfLayer
approach which leaves muted layers in the layer stack but makes it
so they have no opinions.  The Pcp approach drops any refPtr to the
muted layers, likely causing them to be destroyed.  Dropping muted
layers from the layer stack and possibly destroying them violate
our current expectations.

(Internal change: 2354751)