[HdPrman] Add HdPrman MaterialX colorspace handling
This change adds colorspace support to HdPrman. Same as Storm, it
sets the scene referred colorspace to linrec709, and then makes all
non-tex textures get the 'raw' colorspace. This way any colorspace
transformations can happen in the MaterialX generated shaders.

Note that if a material were to author a PreviewSurface diffuseColor
in a particular colorspace without an attached nodegraph, this would
not work since in that case we are not using any MaterialX shaders,
we would be relying more directly on the PxrSurface and so no
colorspaces transformations would occur.

(Internal change: 2355565)