[Hdsi] Make Render Settings Prim Path stored on the Scene Globals Scene Index std::optional
In Hydra 2.0 the render settings path from the stage's metadata is
stored on the UsdImaging Stage Scene Index. Then in the the
Scene Globals Scene Index this was overwitten by the value set
via that Scene Indices SetActiveRenderSettingsPrimPath(). This
means that if this function was not called, the value from the
metadata will be overwritten by an empty SdfPath.

This changes the render settings path stored by the Scene Globals
Scene Index to be std::optional so that we make sure to fallback to
the upstream scene index's opion.

Fixes #3367

(Internal change: 2355965)