[hdx] Fix for strict build compilation error in open source when using VFX2023.
I lean towards the reported error (array bounds) being a compiler bug for a few
reasons. 1) It appeared after upgrading to VFX 2023. 2) As far as I can tell,
the error only happens for the first for-loop over highlight modes and not the
second, even though they use the same for-loop conditions, array indices, and
syntax. 3) I haven't been able to recreate the error in the compiler explorer
(although it doesn't currently support the exact setup from what I can tell
i.e. gcc 11.2.1/vfx2023). Simple workarounds like unrolling for-loops or
replacing the problematic for-loop entirely with hard-coded statements also
fixed the issue. I'm guessing the size_t typing is allowing the compiler to
handle bounds checking somewhat differently and avoid the spurious error. One
theory is that since the first for-loop is setting all the elements to the same
value unconditionally while the second is not, the compiler might be
optimizing the first for-loop into a memset (__builtin_memset) that isn't
working correctly with the 32-bit index values.

(Internal change: 2356294)