Fixing a bug where UsdPrim::ComputeExpandedPrimIndex was not computing the prim index in USD mode, causing it to ignore instancing restrictions among other things.
The underlying issue was that PcpCache::GetPrimIndexInputs wasn't adding
the USD mode flag to the returned inputs. PcpPrimIndex-internal uses of
of GetPrimIndexInputs were adding the USD mode flag to them manually but
the external uses of it (namely ComputeExpandedPrimIndex) did not know
to do this.

In the end I refactored the GetPrimIndexInputs function to return a
const ref result that always includes the USD mode flag and made clients
that want to change the inputs have to explicitly make a copy of the
inputs. I also wrapped PcpPrimIndex::IsUsd to python for easy testing.

Fixes #3526

(Internal change: 2356795)