[hdPrman] Get nodeType/nodedef name from the MaterialX document instead of the material network
HdPrman currently looks to the nodeInterface to find the nodeType,
however, if the nodedef name has changed between versions of
MaterialX and we relied on MaterialX's document upgrade process
used in HdMtlx, we would not get the correct nodeType. This
change instead looks to the MaterialX document for the nodeType.
Since the document will already have been upgraded at this point,
it wil contain the correct updated nodedef/nodeType name.

(Internal change: 2356901)